+++ frontend/src/Routes.tsx
@@ -4,7 +4,7 @@ import AuthLogout from "@/components/auth/AuthLogout"
 import ControllerBindingsModal from "@/components/modals/ControllerBindingsModal"
 import ProjectFormModal from "@/components/modals/ProjectFormModal"
 import SplashLogo from "@/components/SplashLogo"
-import AuthCallback from "@/components/user/AuthCallback"
+import AuthCallback from "@/components/auth/AuthCallback"
 import ConfigListPage from "@/pages/ConfigList"
 import Dashboard from "@/pages/Dashboard"
 import Plain from "@/pages/Plain"
